Relocated the TM endpoints to exist within the controller package Relocated the DTO objects to exist in a data package
