+++ src/main/java/uk/gov/ons/fwmt/legacy_gateway/controller/tm_endpoint/WebConfig.java
@@ -1,4 +1,4 @@
-package uk.gov.ons.fwmt.legacy_gateway.tm.endpoint;
+package uk.gov.ons.fwmt.legacy_gateway.controller.tm_endpoint;
 
 import org.springframework.boot.web.servlet.ServletRegistrationBean;
 import org.springframework.context.ApplicationContext;
